# tdd_implementation/step2_api_unification_simple.py
"""
TDD测试驱动开发：认知陷阱平台后端重构
第二步：实现API端点统一 - 简化版
"""
import sys
import os

def implement_unified_api_endpoints():
    """实现统一API端点 - 通过直接字符串替换实现"""
    print("正在实现统一API端点（简化版）...")
    
    # 读取当前的start.py文件
    with open("api-server/start.py", "r", encoding="utf-8") as f:
        content = f.read()
    
    # 检查是否已实现难度参数
    if 'difficulty: str = Query("auto"' in content:
        print("✓ create_game_session函数已包含difficulty参数")
    else:
        # 需要更新create_game_session函数定义
        # 替换函数签名
        old_func_sig = 'async def create_game_session(scenario_id: str = Query(..., alias="scenario_id")'
        
        if old_func_sig in content:
            new_func_sig = 'async def create_game_session(scenario_id: str = Query(..., alias="scenario_id"), difficulty: str = Query("auto", description="难度级别: beginner, intermediate, advanced, 或 auto")'
            updated_content = content.replace(old_func_sig, new_func_sig)
            
            # 更新函数体处理逻辑
            if 'selected_scenario = scenario.copy()' in updated_content:
                # 已经有难度处理逻辑，跳过
                print("✓ 检测到现有的难度处理逻辑")
            else:
                # 添加难度处理逻辑到函数体中
                # 找到函数体开始的地方
                if 'selected_scenario = scenario.copy()' not in content and 'if difficulty != "auto"' not in content:
                    # 查找函数实现部分并添加难度处理
                    function_start = updated_content.find('async def create_game_session(')
                    if function_start != -1:
                        # 找到函数主体部分
                        body_start = updated_content.find(':', function_start)
                        if body_start != -1:
                            colon_pos = body_start + 1
                            # 查找下一行开始
                            next_line = updated_content.find('\n', colon_pos)
                            if next_line != -1:
                                # 插入难度处理逻辑
                                indent = '    '  # 4空格缩进
                                difficulty_logic = f'''
{indent}    # 根据难度参数调整场景
{indent}    selected_scenario = scenario.copy()
{indent}
{indent}    if difficulty != "auto":
{indent}        # 如果指定了具体难度，查找对应的高级挑战内容
{indent}        if difficulty != scenario["difficulty"]:
{indent}            # 在高级挑战中查找匹配难度的挑战
{indent}            matching_challenge = None
{indent}            if "advancedChallenges" in scenario:
{indent}                for challenge in scenario["advancedChallenges"]:
{indent}                    if challenge["difficulty"] == difficulty:
{indent}                        matching_challenge = challenge
{indent}                        break
{indent}
{indent}            if matching_challenge:
{indent}                # 用高级挑战的信息更新场景
{indent}                selected_scenario["name"] = f"{{scenario['name']}} - {{matching_challenge['title']}}"
{indent}                selected_scenario["description"] = matching_challenge["description"]
{indent}                selected_scenario["targetBiases"] = matching_challenge["cognitiveBiases"]
{indent}                selected_scenario["cognitiveBias"] = ", ".join(matching_challenge["cognitiveBiases"])
'''
                                
                                # 在函数体开始处插入难度处理逻辑
                                new_content = updated_content[:next_line+1] + difficulty_logic + updated_content[next_line+1:]
                                updated_content = new_content
            
            content = updated_content
            print("✓ 已更新create_game_session函数以包含difficulty参数和处理逻辑")
        else:
            print("✓ 已存在难度处理逻辑")
    
    # 更新execute_turn函数以支持难度参数
    if 'difficulty = session.get("difficulty", "beginner")' in content:
        print("✓ execute_turn函数已支持难度参数")
    else:
        # 需要更新execute_turn函数
        old_execute_turn = 'def execute_turn(game_id: str, decisions: Dict[str, Any]):'
        if old_execute_turn in content:
            # 更新execute_real_logic调用以传递difficulty参数
            old_call = 'new_state = execute_real_logic(scenario_id, current_state, decisions)'
            new_call = 'new_state = execute_real_logic(scenario_id, current_state, decisions, difficulty=difficulty)'
            
            if old_call in content:
                content = content.replace(old_call, new_call)
                
                # 更新generate_real_feedback调用以传递difficulty参数
                old_feedback_call = 'feedback = generate_real_feedback(scenario_id, decisions, current_state, new_state)'
                new_feedback_call = 'feedback = generate_real_feedback(scenario_id, decisions, current_state, new_state, difficulty=difficulty)'
                
                if old_feedback_call in content:
                    content = content.replace(old_feedback_call, new_feedback_call)
                
                # 添加获取difficulty的代码行
                if 'difficulty = session.get(' not in content:
                    # 查找execute_turn函数开始处并插入difficulty获取
                    func_start = content.find('def execute_turn(game_id: str, decisions: Dict[str, Any]):')
                    if func_start != -1:
                        # 找到函数体
                        colon_pos = content.find(':', func_start)
                        if colon_pos != -1:
                            next_newline = content.find('\n', colon_pos)
                            if next_newline != -1:
                                # 插入difficulty获取逻辑
                                indent = '    '  # 4空格缩进
                                difficulty_extract = f'''
{indent}    difficulty = session.get("difficulty", "beginner")  # 获取难度级别
'''
                                content = content[:next_newline+1] + difficulty_extract + content[next_newline+1:]
        
        print("✓ 已更新execute_turn函数以支持difficulty参数")
    
    # 写入更新后的内容
    with open("api-server/start.py", "w", encoding="utf-8") as f:
        f.write(content)
    
    print("✓ 统一API端点实现完成")
    return True

def implement_business_logic_difficulty_support():
    """实现业务逻辑的难度支持"""
    print("正在实现业务逻辑的难度支持...")
    
    # 检查execute_real_logic函数是否支持difficulty参数
    with open("api-server/start.py", "r", encoding="utf-8") as f:
        content = f.read()
    
    # 检查函数定义
    if 'def execute_real_logic(' in content:
        # 检查是否已经有difficulty参数
        import re
        # 查找函数定义行
        func_def_match = re.search(r'def execute_real_logic\(([^)]*)\)', content)
        if func_def_match:
            params = func_def_match.group(1)
            if 'difficulty' not in params:
                # 更新函数定义以添加difficulty参数
                old_def = 'def execute_real_logic(scenario_id: str, current_state: Dict, decisions: Dict)'
                new_def = 'def execute_real_logic(scenario_id: str, current_state: Dict, decisions: Dict, difficulty: str = "beginner")'
                
                if old_def in content:
                    content = content.replace(old_def, new_def)
                
                print("✓ 已更新execute_real_logic函数签名以支持difficulty参数")
            else:
                print("✓ execute_real_logic函数签名已支持difficulty参数")
        
        # 检查函数体实现
        lines = content.split('\n')
        new_lines = []
        in_func = False
        func_indent = None
        
        for i, line in enumerate(lines):
            if 'def execute_real_logic(' in line:
                in_func = True
                # 找到函数体的缩进级别
                func_indent = len(line) - len(line.lstrip())
                new_lines.append(line)
            elif in_func and line.strip() == '' and len(line.lstrip()) <= func_indent:
                # 函数结束了
                in_func = False
                func_indent = None
                new_lines.append(line)
            elif in_func:
                # 檢查是否需要修改特定的计算逻辑
                stripped = line.strip()
                new_lines.append(line)
            else:
                new_lines.append(line)
    
    # 检查generate_real_feedback函数
    if 'def generate_real_feedback(' in content:
        import re
        func_def_match = re.search(r'def generate_real_feedback\(([^)]*)\)', content)
        if func_def_match:
            params = func_def_match.group(1)
            if 'difficulty' not in params:
                # 更新函数定义
                old_def = 'def generate_real_feedback(scenario_id: str, decisions: Dict, old_state: Dict, new_state: Dict)'
                new_def = 'def generate_real_feedback(scenario_id: str, decisions: Dict, old_state: Dict, new_state: Dict, difficulty: str = "beginner")'
                
                if old_def in content:
                    content = content.replace(old_def, new_def)
                
                print("✓ 已更新generate_real_feedback函数签名以支持difficulty参数")
            else:
                print("✓ generate_real_feedback函数签名已支持difficulty参数")
    
    # 写入更新后的内容
    with open("api-server/start.py", "w", encoding="utf-8") as f:
        f.write(content)
    
    print("✓ 业务逻辑难度支持实现完成")
    return True

def verify_implementation():
    """验证API端点统一实现结果"""
    print("正在验证API端点统一实现结果...")
    
    # 重新读取文件验证修改
    with open("api-server/start.py", "r", encoding="utf-8") as f:
        content = f.read()
    
    # 检查create_game_session是否包含difficulty参数
    if 'difficulty: str = Query(' in content and 'create_game_session' in content:
        print("✓ create_game_session函数包含difficulty参数")
    else:
        print("! 警告: create_game_session函数可能缺少difficulty参数")
    
    # 检查execute_real_logic是否包含difficulty参数
    if 'def execute_real_logic(' in content and 'difficulty: str = "beginner"' in content:
        print("✓ execute_real_logic函数支持difficulty参数")
    else:
        print("! 警告: execute_real_logic函数可能缺少difficulty参数")
    
    # 检查generate_real_feedback是否包含difficulty参数
    if 'def generate_real_feedback(' in content and 'difficulty: str = "beginner"' in content:
        print("✓ generate_real_feedback函数支持difficulty参数")
    else:
        print("! 警告: generate_real_feedback函数可能缺少difficulty参数")
    
    # 检查端点是否正常
    if '@app.post("/scenarios/create_game_session"' in content:
        print("✓ create_game_session端点已定义")
    else:
        raise AssertionError("create_game_session端点未定义")
    
    print("✓ API端点统一实现验证完成")
    return True

if __name__ == "__main__":
    print("开始TDD实施: API端点统一（简化版）\n")
    
    try:
        # 实施API端点统一
        implement_unified_api_endpoints()
        implement_business_logic_difficulty_support()
        
        # 验证实施结果
        verify_implementation()
        
        print("\n🎉 API端点统一TDD实施成功完成!")
        print("现在可以继续下一步: 前端界面开发")
        
    except Exception as e:
        print(f"\n❌ 实施失败: {e}")
        import traceback
        traceback.print_exc()