# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"认知陷阱平台API","description":"提供认知陷阱场景、游戏会话和分析服务，使用真实的逻辑实现","version":"1.0.0"},"paths":{"/api/exponential/questions":{"get":{"tags":["cognitive_tests"],"summary":"Get Exponential Questions","description":"获取指数增长相关的测试问题","operationId":"get_exponential_questions_api_exponential_questions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/exponential/calculate/exponential":{"post":{"tags":["cognitive_tests"],"summary":"Calculate Exponential Endpoint","description":"计算指数增长结果","operationId":"calculate_exponential_endpoint_api_exponential_calculate_exponential_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExponentialRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/exponential/calculate/granary":{"post":{"tags":["cognitive_tests"],"summary":"Calculate Granary Problem","description":"计算米粒问题（2^200粒米的重量和体积）","operationId":"calculate_granary_problem_api_exponential_calculate_granary_post","parameters":[{"name":"grains_per_unit","in":"query","required":false,"schema":{"type":"integer","description":"每单位的米粒数","default":1,"title":"Grains Per Unit"},"description":"每单位的米粒数"},{"name":"rice_weight_per_grain_g","in":"query","required":false,"schema":{"type":"number","description":"每粒米的重量(克)","default":0.02,"title":"Rice Weight Per Grain G"},"description":"每粒米的重量(克)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/exponential/calculate/rabbit-growth":{"post":{"tags":["cognitive_tests"],"summary":"Calculate Rabbit Growth","description":"模拟兔子增长 - 从10只兔子，每年翻5倍，11年后会有多少只","operationId":"calculate_rabbit_growth_api_exponential_calculate_rabbit_growth_post","parameters":[{"name":"starting_rabbits","in":"query","required":false,"schema":{"type":"integer","description":"起始兔子数量","default":10,"title":"Starting Rabbits"},"description":"起始兔子数量"},{"name":"years","in":"query","required":false,"schema":{"type":"integer","description":"年数","default":11,"title":"Years"},"description":"年数"},{"name":"growth_multiplier","in":"query","required":false,"schema":{"type":"integer","description":"每年增长倍数","default":5,"title":"Growth Multiplier"},"description":"每年增长倍数"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/exponential/calculate/compare-linear-exponential":{"post":{"tags":["cognitive_tests"],"summary":"Compare Linear Exponential","description":"比较线性增长与指数增长的差异","operationId":"compare_linear_exponential_api_exponential_calculate_compare_linear_exponential_post","parameters":[{"name":"initial_amount","in":"query","required":true,"schema":{"type":"number","description":"初始金额","title":"Initial Amount"},"description":"初始金额"},{"name":"rate_percent","in":"query","required":true,"schema":{"type":"number","description":"增长率百分比","title":"Rate Percent"},"description":"增长率百分比"},{"name":"time_periods","in":"query","required":true,"schema":{"type":"integer","description":"时间周期数","title":"Time Periods"},"description":"时间周期数"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/compound/questions":{"get":{"tags":["cognitive_tests"],"summary":"Get Compound Questions","description":"获取复利相关的测试问题","operationId":"get_compound_questions_api_compound_questions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/compound/calculate/interest":{"post":{"tags":["cognitive_tests"],"summary":"Calculate Compound Interest Endpoint","description":"计算复利结果","operationId":"calculate_compound_interest_endpoint_api_compound_calculate_interest_post","parameters":[{"name":"principal","in":"query","required":true,"schema":{"type":"number","description":"本金","title":"Principal"},"description":"本金"},{"name":"annual_rate","in":"query","required":true,"schema":{"type":"number","description":"年利率（%）","title":"Annual Rate"},"description":"年利率（%）"},{"name":"time_years","in":"query","required":true,"schema":{"type":"integer","description":"年数","title":"Time Years"},"description":"年数"},{"name":"compounding_frequency","in":"query","required":false,"schema":{"type":"integer","description":"复利频率（次/年）","default":1,"title":"Compounding Frequency"},"description":"复利频率（次/年）"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/historical/scenarios":{"get":{"tags":["cognitive_tests"],"summary":"Get Historical Scenarios","description":"获取历史决策案例","operationId":"get_historical_scenarios_api_historical_scenarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/game/scenarios":{"get":{"tags":["cognitive_tests"],"summary":"Get Game Scenarios","description":"获取推理游戏场景","operationId":"get_game_scenarios_api_game_scenarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/results/submit":{"post":{"tags":["cognitive_tests"],"summary":"Submit User Response","description":"提交用户响应结果","operationId":"submit_user_response_api_results_submit_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/results/{user_id}/{session_id}":{"get":{"tags":["cognitive_tests"],"summary":"Get Session Results","description":"获取特定用户的会话结果","operationId":"get_session_results_api_results__user_id___session_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/explanations/{bias_type}":{"get":{"tags":["cognitive_tests"],"summary":"Get Bias Explanation","description":"获取特定认知偏差的详细解释","operationId":"get_bias_explanation_api_explanations__bias_type__get","parameters":[{"name":"bias_type","in":"path","required":true,"schema":{"type":"string","title":"Bias Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"API根端点","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/scenarios/":{"get":{"summary":"Get Scenarios","description":"获取所有认知陷阱场景","operationId":"get_scenarios_scenarios__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/scenarios/{scenario_id}":{"get":{"summary":"Get Scenario","description":"获取特定场景详情","operationId":"get_scenario_scenarios__scenario_id__get","parameters":[{"name":"scenario_id","in":"path","required":true,"schema":{"type":"string","title":"Scenario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/scenarios/create_game_session":{"post":{"summary":"Create Game Session","description":"创建游戏会话","operationId":"create_game_session_scenarios_create_game_session_post","parameters":[{"name":"scenario_id","in":"query","required":true,"schema":{"type":"string","title":"Scenario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/scenarios/{game_id}/turn":{"post":{"summary":"Execute Turn","description":"执行游戏回合（真实逻辑实现）","operationId":"execute_turn_scenarios__game_id__turn_post","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Decisions"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ExponentialRequest":{"properties":{"base":{"type":"number","title":"Base"},"exponent":{"type":"integer","title":"Exponent"}},"type":"object","required":["base","exponent"],"title":"ExponentialRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}